Fixed warning about float to LONGLONG conversion and removed obsolete comment.
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@1002 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ textbox.cpp
@@ -289,9 +289,7 @@ void Textbox::RenderSetup(const RenderDevice* _pd3dDevice)
 
     CY size;
     m_pIFontPlay->get_Size(&size);
-    // I choose 912 because that was the original playing size I tested with,
-    // and this way I don't have to change my tables
-    size.int64 = size.int64 / 1.5f*ppin3d->m_dwRenderHeight * ppin3d->m_dwRenderWidth;
+    size.int64 = (LONGLONG)(size.int64 / 1.5f*ppin3d->m_dwRenderHeight * ppin3d->m_dwRenderWidth);
     m_pIFontPlay->put_Size(size);
 
     RenderText();
